fix: Add padding to PostCard Fixes #40
